+++ src/main/java/pe/dinnra_web/sistema_gestion/api/exceptions/PositionNotFoundException.java
@@ -1,0 +1,7 @@
+package pe.dinnra_web.sistema_gestion.api.exceptions;
+
+public class PositionNotFoundException extends RuntimeException {
+    public PositionNotFoundException(String message) {
+        super(message);
+    }
+}
