+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/ContactServiceType.java
@@ -1,0 +1,8 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum ContactServiceType {
+    RECLAMO_RESERVA,
+    RECLAMO_EMPLEADO_SERVICIOS,
+    RECLAMO_SERVICIOS,
+    OTROS
+}
