+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/EmployeeStatus.java
@@ -1,0 +1,6 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum EmployeeStatus {
+    DISPONIBLE,
+    OCUPADO
+}
