+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/Gender.java
@@ -1,0 +1,7 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum Gender {
+    MASCULINO,
+    FEMENINO,
+    OTRO
+}
