+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/ReservationStatus.java
@@ -1,0 +1,8 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum ReservationStatus {
+    REGISTRADA,
+    ACTIVA,
+    CERRADA,
+    CANCELADA
+}
