+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/RoomStatus.java
@@ -1,0 +1,7 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum RoomStatus {
+    DISPONIBLE,
+    OCUPADO,
+    MANTENIMIENTO
+}
