+++ src/main/java/pe/dinnra_web/sistema_gestion/api/model/enums/ServiceRegistrationLogStatus.java
@@ -1,0 +1,8 @@
+package pe.dinnra_web.sistema_gestion.api.model.enums;
+
+public enum ServiceRegistrationLogStatus {
+    PENDIENTE,
+    EN_PROGRESO,
+    COMPLETADO,
+    CANCELADO
+}
