--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/entity/RoomAmenity.java
@@ -1,4 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.entity;
-
-public class RoomAmenity {
-}
