--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/ContactServiceType.java
@@ -1,8 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum ContactServiceType {
-    RECLAMO_RESERVA,
-    RECLAMO_EMPLEADO_SERVICIOS,
-    RECLAMO_SERVICIOS,
-    OTROS
-}
