--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/EmployeeStatus.java
@@ -1,6 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum EmployeeStatus {
-    DISPONIBLE,
-    OCUPADO
-}
