--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/Gender.java
@@ -1,7 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum Gender {
-    MASCULINO,
-    FEMENINO,
-    OTRO
-}
