--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/ReservationStatus.java
@@ -1,8 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum ReservationStatus {
-    REGISTRADA,
-    ACTIVA,
-    CERRADA,
-    CANCELADA
-}
