--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/RoomStatus.java
@@ -1,7 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum RoomStatus {
-    DISPONIBLE,
-    OCUPADO,
-    MANTENIMIENTO
-}
