--- src/main/java/pe/dinnra_web/sistema_gestion/api/models/enums/ServiceRegistrationLogStatus.java
@@ -1,8 +1,0 @@
-package pe.dinnra_web.sistema_gestion.api.models.enums;
-
-public enum ServiceRegistrationLogStatus {
-    PENDIENTE,
-    EN_PROGRESO,
-    COMPLETADO,
-    CANCELADO
-}
