+++ src/main/java/pe/dinnra_web/sistema_gestion/api/repository/PositionRepository.java
@@ -1,0 +1,8 @@
+package pe.dinnra_web.sistema_gestion.api.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import pe.dinnra_web.sistema_gestion.api.model.entity.Position;
+
+public interface PositionRepository extends JpaRepository<Position, Long> {
+
+}
